lambda_processor/requirements.txt
boto3>=1.34
psycopg2-binary>=2.9
python-json-logger>=2.0
